Move the compatibility check to a separate class
